task 3

+++ preprocessor/3-function_like_macro.h
@@ -1,6 +1,6 @@
 #ifndef __FILE_NAME__
 #define __FILE_NAME__
 #include <stdlib.h>
-#define ABS(x) (abs(x))
+#define ABS(x) (abs((int)x))
 
 #endif
